Add APIs and restructure
Implemented apis for listing and searching for materials.
Restructure all the code and bring some modularity.
Update dependencies.
