+++ requirements.txt
@@ -1,7 +1,8 @@
 autopep8==1.5.7
 blessed==1.17.6
-certifi==2020.12.5
+certifi==2021.5.30
 chardet==4.0.0
+click==8.0.1
 defusedxml==0.7.1
 idna==2.10
 inquirer==2.7.0
@@ -20,5 +21,5 @@ readchar==2.0.1
 requests==2.25.1
 six==1.16.0
 toml==0.10.2
-urllib3==1.26.4
+urllib3==1.26.5
 wcwidth==0.2.5
